+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 ### Added
 - Support for Moodle 4.4
+- Ability to duplicate sections
 ### Fixed
 - Fixed fatal PHP error on Moodle 4.3 introduced by changes in MDL-81610